Register User

+++ base/views.py
@@ -1,12 +1,13 @@
 from urllib.request import Request
 from django.shortcuts import render,redirect
-from django.contrib.auth.models import User
-from django.contrib.messages import constants as messages
-
 from django.db.models import Q
+from django.http import HttpResponse
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate,login,logout
+from django.contrib import messages
+from django.contrib.auth.forms import UserCreationForm
 from .models import Room,Topic
-
-
 from .forms import RoomForm
 # Create your views here.
 
@@ -18,16 +19,33 @@ from .forms import RoomForm
 # ]
 
 def loginPage(request):
+    page='login'
+    if request.user.is_authenticated:
+        return redirect('home')
+
+
     if request.method=='POST':
         username=request.POST.get('username')
         password=request.POST.get('password')
-
         try:   
             user=User.objects.get(username=username)
         except:
-           messages.error(request, 'User Does not Exist')
-    context={}
+           messages.error(request, 'User does not exist')
+        user=authenticate(request,username=username,password=password)  
+        if user is not None:
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Credentials Dont Match') 
+    context={page:'login'}
     return render(request,'base/login_register.html',context)
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+def registerPage(request):
+    page='register'
+    return render(request,'base/login_register.html')    
 def home(request):
     #objects-model manager
     q=request.GET.get('q') if request.GET.get('q')!=None else '' #whatever we passed onto the url 
@@ -48,7 +66,7 @@ def room(request,pk):
     context={'room':room}        
     return render(request,'base/room.html',context)   
 
-
+@login_required(login_url='login')
 def createRoom(request):
     form=RoomForm()
     if request.method == 'POST':
@@ -59,10 +77,14 @@ def createRoom(request):
             return redirect('home')
     context={'form':form}
     return render(request,'base/room_form.html',context)
-
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room =Room.objects.get(id=pk)
     form=RoomForm(instance=room)
+
+    if request.user!=room.host:
+        return HttpResponse('You are Tresspassing!!!')
+
     if request.method == 'POST':
         form=RoomForm(request.POST,instance=room)
         if form.is_valid():
@@ -70,9 +92,11 @@ def updateRoom(request,pk):
             return redirect('home')    
     context={'form':form}
     return render(request,'base/room_form.html',context)
-
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room=Room.objects.get(id=pk)
+    if request.user!=room.host:
+        return HttpResponse('You are Tresspassing!!!')
     if request.method == 'POST':
         room.delete()
         return redirect('home')
